svg/animations/onend.svg

<svg xmlns="http://www.w3.org/2000/svg"><title>&apos;onend&apos; event handler content attribute</title><script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharness.js"/><script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharnessreport.js"/><rect width="0" height="100" fill="green"><animate id="anim" fill="freeze" attributeName="visibility" begin="0s" end="2s" onend="document.getElementById('anim2').beginElement()" to="visible"/><set id="anim2" attributeName="width" begin="indefinite" to="100"/></rect><script>async_test(t =&gt; { document.documentElement.setCurrentTime(1.99); let set = document.getElementById(&apos;anim2&apos;); set.addEventListener(&apos;beginEvent&apos;, t.step_func_done(function() { assert_equals(set.parentNode.getBBox().width, 100); })); });</script></svg>
